backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pillow
numpy
pydantic
python-dotenv
orjson